Extract base test class for WorldEdit test cases
